@@ -17,11 +17,15 @@
 
 package org.keycloak.connections.infinispan;
 
-import java.util.List;
+import java.util.Arrays;
 import java.util.concurrent.CompletionStage;
+import java.util.concurrent.Executor;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.stream.Stream;
 
 import org.infinispan.Cache;
 import org.infinispan.client.hotrod.RemoteCache;
+import org.keycloak.models.KeycloakSessionFactory;
 import org.keycloak.provider.Provider;
 
 /**
@@ -51,7 +55,7 @@ public interface InfinispanConnectionProvider extends Provider {
     String ACTION_TOKEN_CACHE = "actionTokens";
     int ACTION_TOKEN_CACHE_DEFAULT_MAX = -1;
     int ACTION_TOKEN_MAX_IDLE_SECONDS = -1;
-    long ACTION_TOKEN_WAKE_UP_INTERVAL_SECONDS = 5 * 60 * 1000l;
+    long ACTION_TOKEN_WAKE_UP_INTERVAL_SECONDS = 5 * 60 * 1000L;
 
     String KEYS_CACHE_NAME = "keys";
     int KEYS_CACHE_DEFAULT_MAX = 1000;
@@ -70,26 +74,19 @@ public interface InfinispanConnectionProvider extends Provider {
     // Constant used as the prefix of the current node if "jboss.node.name" is not configured
     String NODE_PREFIX = "node_";
 
-    String[] ALL_CACHES_NAME = {
+    // list of cache name for local caches (not replicated)
+    String[] LOCAL_CACHE_NAMES = {
             REALM_CACHE_NAME,
             REALM_REVISIONS_CACHE_NAME,
             USER_CACHE_NAME,
             USER_REVISIONS_CACHE_NAME,
-            USER_SESSION_CACHE_NAME,
-            CLIENT_SESSION_CACHE_NAME,
-            OFFLINE_USER_SESSION_CACHE_NAME,
-            OFFLINE_CLIENT_SESSION_CACHE_NAME,
-            LOGIN_FAILURE_CACHE_NAME,
-            AUTHENTICATION_SESSIONS_CACHE_NAME,
-            WORK_CACHE_NAME,
             AUTHORIZATION_CACHE_NAME,
             AUTHORIZATION_REVISIONS_CACHE_NAME,
-            ACTION_TOKEN_CACHE,
             KEYS_CACHE_NAME
     };
 
     // list of cache name which could be defined as distributed or replicated
-    public static List<String> DISTRIBUTED_REPLICATED_CACHE_NAMES = List.of(
+    String[] CLUSTERED_CACHE_NAMES = {
             USER_SESSION_CACHE_NAME,
             CLIENT_SESSION_CACHE_NAME,
             OFFLINE_USER_SESSION_CACHE_NAME,
@@ -97,7 +94,10 @@ public interface InfinispanConnectionProvider extends Provider {
             LOGIN_FAILURE_CACHE_NAME,
             AUTHENTICATION_SESSIONS_CACHE_NAME,
             ACTION_TOKEN_CACHE,
-            WORK_CACHE_NAME);
+            WORK_CACHE_NAME
+    };
+
+    String[] ALL_CACHES_NAME = Stream.concat(Arrays.stream(LOCAL_CACHE_NAMES), Arrays.stream(CLUSTERED_CACHE_NAMES)).toArray(String[]::new);
 
     /**
      *
@@ -136,5 +136,32 @@ public interface InfinispanConnectionProvider extends Provider {
      * @return A {@link CompletionStage} to signal when the operator is completed.
      */
     CompletionStage<Void> migrateToProtostream();
+
+    /**
+     * Returns an executor that will run the given tasks on a blocking thread as required.
+     * <p>
+     * The Infinispan block {@link Executor} is used to execute blocking operation, like I/O.
+     * If Virtual Threads are enabled, this will be an executor with Virtual Threads.
+     *
+     * @param name The name for trace logging purpose.
+     * @return The Infinispan blocking {@link Executor}.
+     */
+    Executor getExecutor(String name);
+
+    /**
+     * @return The Infinispan {@link ScheduledExecutorService}. Long or blocking operations must not be executed directly.
+     */
+    ScheduledExecutorService getScheduledExecutor();
+
+    /**
+     * Syntactic sugar to get a {@link RemoteCache}.
+     *
+     * @see InfinispanConnectionProvider#getRemoteCache(String)
+     */
+    static <K, V> RemoteCache<K, V> getRemoteCache(KeycloakSessionFactory factory, String cacheName) {
+        try (var session = factory.create()) {
+            return session.getProvider(InfinispanConnectionProvider.class).getRemoteCache(cacheName);
+        }
+    }
 
 }
